migrate quantity to float

+++ RecipeManager/Data/Ingredient.cs
@@ -5,7 +5,7 @@
         public int IngredientId { get; set; }
         public int RecipeId { get; set; }
         public string Name { get; set; }
-        public decimal Quantity { get; set; }
+        public float Quantity { get; set; }
         public string Unit { get; set; }
 
     }
